Reorganize physics related files into subdir
